backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
yt-dlp>=2024.12.0
openai>=1.50.0
youtube-transcript-api>=0.6.3
aiosqlite>=0.20.0
pydub>=0.25.1
python-dotenv>=1.0.1
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
